refactor(liquid): Split out into separate file
